from ic methiod for tumnable transmon and update in resonators
The calculation using the Cp factor corrections for 1/4 resonaytors (C* = 1 Cp) to account fo a single node affecting the capacitance makes a model that does not fit well the measured resonators. It is possible that we take into consideration the capacitance of onlyt one node, but twice because of the double roundtrip of the wave. In practice both equations becomne equivalent, but we keep them separated so that in the future we can calculate the precise contribution of Ck based on its position along the resonator (Cg is always 1)

Added impedance and capaciatnce calculation utils to the utils module